refactor: split analysis API into audio/analysis.py + add E2E levels tests (#280)
Review follow-up for PR #281:

1. Move NoiseAnalysis and analyze_noise_samples from noise_gate.py to
   a new audio/analysis.py module. The calibration API is a pure
   function with no dependency on NoiseGate internals or the JIT
   _process_loop, so separating it clarifies responsibility and makes
   it easier to reuse from the GUI without pulling in numba state.

2. Add behavior tests for levels --json / --duration. The prior tests
   only verified --help exposed the options. TestLevelsBehavior now
   monkeypatches MicrophoneSource with a fake that returns constant
   -60 dB noise, then exercises cli.main(["levels", "--duration", ...,
   "--json"]) to verify:
   - JSON stdout is parseable
   - All NoiseAnalysis fields are present
   - suggested_threshold_db == noise_peak_db + 10
   - --duration auto-stops within a bounded wall-clock time

Tests moved: TestAnalyzeNoiseSamples → tests/audio/test_analysis.py.
33/33 passed.
